# uas/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is the API documentation for My API.",
        "termsOfService": "https://dev.test/terms",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://dev.test",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Tampilkan daftar buku",
                "description": "Mengambil daftar semua buku dengan paginasi.",
                "operationId": "getBooksList",
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Buat buku baru",
                "description": "Menyimpan data buku baru ke database.",
                "operationId": "storeBook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Buku berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Data tidak valid"
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Tampilkan detail buku",
                "description": "Mengambil data satu buku berdasarkan ID.",
                "operationId": "getBookById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Buku",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Books"
                ],
                "summary": "Perbarui data buku",
                "description": "Memperbarui data buku yang sudah ada di database.",
                "operationId": "updateBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Buku",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Buku berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    },
                    "422": {
                        "description": "Data tidak valid"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Hapus data buku",
                "description": "Menghapus data buku dari database.",
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Buku",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Buku berhasil dihapus"
                    },
                    "404": {
                        "description": "Buku tidak ditemukan"
                    }
                }
            }
        },
        "/api/tests": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get all tests",
                "operationId": "5f539f69bb1d910182eb35136c5baa3a",
                "responses": {
                    "200": {
                        "description": "List of tests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "List of tests retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Test"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Create a new test",
                "operationId": "7728a2f3dd87105d6d617df9a3f231d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Test description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Test created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Test created successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Test"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/tests/{id}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get a specific test",
                "operationId": "7e3d8428f4df82c6d4ee7bd1d4d2128c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Test retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Test"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Test not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tests"
                ],
                "summary": "Update a specific test",
                "operationId": "ca1490751234c723e0a4a708afe16dd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Test Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Test updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Test"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Test not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Delete a specific test",
                "operationId": "92f76a68796679554c71a4659f62b296",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the test",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test deleted successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Test deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Test not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Book",
                "required": [
                    "id",
                    "title",
                    "author",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "description": "Primary Key",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Judul Buku",
                        "type": "string"
                    },
                    "author": {
                        "description": "Penulis Buku",
                        "type": "string"
                    },
                    "stock": {
                        "description": "Jumlah stok buku yang tersedia",
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreBookRequest": {
                "title": "Store Book Request",
                "required": [
                    "title",
                    "author",
                    "stock"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Belajar Laravel dari Dasar"
                    },
                    "author": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "stock": {
                        "type": "integer",
                        "example": 15
                    }
                },
                "type": "object"
            },
            "Test": {
                "title": "Test",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Sample Test"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is a test record."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "API Endpoints untuk mengelola data buku"
        },
        {
            "name": "Tests",
            "description": "Tests"
        }
    ]
}